media-fonts/hkscs-ming: remove package, close #184.
This ebuild was left untouched since 12 May, 2004. It is a living fossil,
a witness to history about the creation of the Gentoo Taiwan, and the
grand unification between the gentoo-china (Mainland) and gentoo-tw
communities.

The link is dead, the approaches to install font was the ancient one
(before font.ebuild appears!). It is time to let it come to an end.