media-fonts/lxgwwenkai: bump to version 1.222
Simplified the ebuild because the upstream now provides a tarball
that includes all the ttf files. The tarball name is different
from the previous naming convention though. To avoid disturbing
the existing users of this font, the necessary steps to handle
the new naming convention are added to the ebuild too.