net-im/wemeet: add 3.19.1.400
Also add loong support. It is not currently keyworded though, because
the Chromium sandbox of qtwebengine:5 is crashing right now on statx
calls, preventing successful login.

While at it, also streamline the unbundling process of libraries. It was
discovered that the bundled Qt has a different ABI than the system one,
even on amd64, so the bundled Qt is kept and the RDEPEND got overhauled
because of that.

Signed-off-by: WANG Xuerui <xen0n@gentoo.org>